Added some Server to Client Instructions

+++ FurcLib/Net/Proxy/BaseProxy.cs
@@ -4,30 +4,60 @@
  * (Oct 27,2009) Squizzle, Added NetMessage, delegates, and NetProxy wrapper class.
  * (July 26, 2011) Gerolkae, added setting.ini switch for proxy.ini
  * (Mar 12,2014,0.2.12) Gerolkae, Adapted Paths to work with a Supplied path
+ * (March 2017, 2.1.x) Gerolkae, Furcadia V31 support.
  */
 
 using Furcadia.IO;
+using Furcadia.Net.Utils;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 
-//using System.ComponentModel;
 using System.Net;
 using System.Net.NetworkInformation;
 using System.Net.Sockets;
+using System.Threading;
 using System.Timers;
-
-/// using System.Timers.Timer;
 
 namespace Furcadia.Net
 {
     /// <summary>
     /// Base proxy class to connect to Furcadia game server.
+    /// <para>
+    /// Currently only a furcadia character.ini file is supported.
+    /// </para>
     /// </summary>
+    /// <remarks>
+    /// here we provide the base proxy class. Our Purpose is to connect to the Furcadia Game server
+    /// with the raw sever to client instructions as Line Feed "\n" delimited text.
+    /// <para>
+    /// We modify furcadia's settings.ini file Firewall settings and disable the SSL/TLS connection
+    /// for the Furcadia Client.
+    /// <para>
+    /// Once we have connected we attempt to restore the setting.ini file to its previous status
+    /// after a ten-second delay to allow a Furcadia News update.
+    /// </para>
+    /// </para>
+    /// </remarks>
     public class BaseProxy : IDisposable
     {
         #region Event Handling
+
+        /// <summary>
+        /// This is triggered when the Client sends data to the server. Expects a return value.
+        /// </summary>
+        public EventHandler ClientData2;
+
+        /// <summary>
+        /// Furcadia Client connection status event
+        /// </summary>
+        public EventHandler<NetClientArgs> OnClientStatusChanged;
+
+        /// <summary>
+        /// Furcadia Server Connection status event
+        /// </summary>
+        public EventHandler<NetServerArgs> OnServerStatusChanged;
 
         /// <summary>
         /// </summary>
@@ -47,6 +77,14 @@ namespace Furcadia.Net
 
         /// <summary>
         /// </summary>
+        /// <param name="e">
+        /// Executing Exeception
+        /// </param>
+        /// <param name="o">
+        /// Object
+        /// </param>
+        /// <param name="n">
+        /// </param>
         public delegate void ErrorEventHandler(Exception e, object o, string n);
 
         /// <summary>
@@ -54,26 +92,7 @@ namespace Furcadia.Net
         /// </summary>
         public event DataEventHandler ClientData;
 
-        /// <summary>
-        /// This is triggered when the Client sends data to the server. Expects a return value.
-        /// </summary>
-        public event DataEventHandler2 ClientData2;
-
-        /// <summary>
-        ///This is triggered when the Client Disconnects
-        /// </summary>
-        public event ActionDelegate ClientDisConnected;
-
-        /// <summary>
-        /// This is triggered when the user has exited/logoff Furcadia and the Furcadia client is closed.
-        /// </summary>
-        public event ActionDelegate ClientExited;
-
         //public delegate void ErrorEventHandler(Exception e);
-        /// <summary>
-        ///This is triggered when the
-        /// </summary>
-        public event ActionDelegate Connected;
 
         /// <summary>
         /// This is triggered when a handled Exception is thrown.
@@ -81,7 +100,8 @@ namespace Furcadia.Net
         public event ErrorEventHandler Error;
 
         /// <summary>
-        /// This is triggered when t client is closed.
+        /// This is Triggered when Furcadia "settings.ini" is restored to previous settings before
+        /// proxy was used.
         /// </summary>
         public event ActionDelegate FurcSettingsRestored;
 
@@ -95,21 +115,24 @@ namespace Furcadia.Net
         /// </summary>
         public event DataEventHandler2 ServerData2;
 
-        /// <summary>
-        ///This is triggered when the Server Disconnects
-        /// </summary>
-        public event ActionDelegate ServerDisConnected;
-
         #endregion Event Handling
 
         #region Private Declarations
 
         /// <summary>
+        /// Local host TCP Port
         /// </summary>
+        [CLSCompliant(false)]
         public static int _lport = 6700;
 
         private static bool _StandAloneMode = false, Clientflag = true;
 
+        /// <summary>
+        /// Furcadia Settings to backup,
+        /// <para>
+        /// Proxy/Firewall and SSL/TLS settings.
+        /// </para>
+        /// </summary>
         private static string[] BackupSettings;
 
         /// <summary>
@@ -118,15 +141,30 @@ namespace Furcadia.Net
         private static int BUFFER_CAP = 2048;
 
         private static int ENCODE_PAGE = 1252;
+
         private static System.Timers.Timer NewsTimer;
+
         private IPEndPoint _endpoint;
+
         private string _proc = "Furcadia.exe", _procpath, _procCMD = "-pick";
+
         private int _procID;
+
         private string _ServerLeftOvers;
+
         private bool CConnected = false;
+
         private TcpClient client = new TcpClient();
+
         private byte[] clientBuffer = new byte[BUFFER_CAP], serverBuffer = new byte[BUFFER_CAP];
+
         private string clientBuild, serverBuild;
+
+        /// <summary>
+        /// Protect Furcadia Settings.ini from multiple instances attempting to modify settings.
+        /// </summary>
+        private Mutex FurcMutex;
+
         private TcpListener listen;
         private TcpClient server;
 
@@ -277,6 +315,9 @@ namespace Furcadia.Net
 
         #region Properties
 
+        /// <summary>
+        /// Is the Furcadia Client connected to our proxy?
+        /// </summary>
         public bool IsClientConnected
         {
             get
@@ -358,6 +399,8 @@ namespace Furcadia.Net
 
         #region Public Static Properties
 
+        /// <summary>
+        /// </summary>
         public static int BufferCapacity
         {
             get
@@ -366,6 +409,8 @@ namespace Furcadia.Net
             }
         }
 
+        /// <summary>
+        /// </summary>
         public static int EncoderPage
         {
             get
@@ -410,7 +455,7 @@ namespace Furcadia.Net
         /// <summary>
         /// Connects to the Furcadia Server and starts the mini proxy.
         /// </summary>
-        public void Connect()
+        public virtual void Connect()
         {
             try
             {
@@ -451,24 +496,35 @@ namespace Furcadia.Net
                     }
                     else throw new NetProxyException("Port " + _lport.ToString() + " is being used.");
                 }
+
+                string userName = System.Security.Principal.WindowsIdentity.GetCurrent().Name;
+                FurcMutex = new Mutex(false, "FurcadiaFrameWork" + userName);
+                if (!FurcMutex.WaitOne(100, false))
+                {
+                    Console.WriteLine("Another instance of Furcadia Framework is running.");
+                    // Abort Connection Process
+
+                    return;
+                }
                 string proxyIni = "localhost " + _lport.ToString();
                 FurcPath = new Paths(_procpath);
-                /// UAC Perms Needed to Create proxy.ini Win 7 Change your UAC Level or add file
-                /// create Permissions to [%program files%/furcadia] Maybe Do this at install
+                // UAC Perms Needed to Create proxy.ini Win 7 Change your UAC Level or add file
+                // create Permissions to [%program files%/furcadia] Maybe Do this at install
                 string fIni = Path.Combine(FurcPath.GetInstallPath(), "proxy.ini");
-                /// Check proxy.ini if it exoists then use it
-                ///
-                /// otherwise use settings.ini to avoid UAC issues on %Program Files%
+                // Check proxy.ini if it exists then use it
+
+                // otherwise use settings.ini to avoid UAC issues on %Program Files%
                 if (File.Exists(proxyIni))
                     File.Delete(proxyIni);
                 UseProxyIni = false;
                 BackupSettings = Settings.InitializeFurcadiaSettings(FurcPath.GetLocalSettingsPath());
+
                 //Run
                 if (string.IsNullOrEmpty(ProcessPath)) ProcessPath = FurcPath.GetInstallPath();
                 //check ProcessPath is not a directory
                 if (!Directory.Exists(ProcessPath)) throw new NetProxyException("Process path not found.");
                 if (!File.Exists(Path.Combine(ProcessPath, Process))) throw new NetProxyException("Client executable '" + Process + "' not found.");
-                Process proc = new System.Diagnostics.Process(); //= System.Diagnostics.Process.Start(Process,ProcessCMD );
+                Process proc = new System.Diagnostics.Process();
                 proc.EnableRaisingEvents = true;
                 proc.StartInfo.FileName = Process;
                 proc.StartInfo.Arguments = ProcessCMD;
@@ -476,11 +532,9 @@ namespace Furcadia.Net
                 proc.Start();
                 proc.Exited += delegate
                 {
-                    if (ClientExited != null)
-                    {
-                        CConnected = false;
-                        ClientExited();
-                    }
+                    NetClientArgs Args = new NetClientArgs(Phase.ClientDisconnected);
+                    OnClientStatusChanged(this, Args);
+                    CConnected = false;
                 };
                 ProcID = proc.Id;
                 CConnected = true;
@@ -547,7 +601,7 @@ namespace Furcadia.Net
             catch (Exception e)
             {
                 if (Error != null)
-                    Error?.Invoke(e, this, "Kill()");
+                    Error?.Invoke(e, this, "Dispose()");
             }
         }
 
@@ -610,7 +664,8 @@ namespace Furcadia.Net
             {
                 if (Error != null)
                     Error?.Invoke(e, this, "SendServer");
-                ServerDisConnected?.Invoke();
+                NetServerArgs Args = new NetServerArgs(Phase.Disconnected);
+                OnServerStatusChanged(this, Args);
             }
         }
 
@@ -637,6 +692,7 @@ namespace Furcadia.Net
             {
                 if (BackupSettings != null)
                     Settings.RestoreFurcadiaSettings(BackupSettings);
+                if (FurcMutex != null) FurcMutex.Dispose();
                 // Free other state (managed objects).
             }
             // Free your own state (unmanaged objects). Set large fields to null.
@@ -673,22 +729,21 @@ namespace Furcadia.Net
                 client.GetStream().BeginRead(clientBuffer, 0, clientBuffer.Length, new AsyncCallback(GetClientData), client);
                 server.GetStream().BeginRead(serverBuffer, 0, serverBuffer.Length, new AsyncCallback(GetServerData), server);
 
-                if (Connected != null)
+                NetClientArgs Args = new NetClientArgs(Phase.SettingsSet);
+                OnClientStatusChanged(this, Args);
+
+                Clientflag = true;
+                // Delete proxy.ini or restore settings.ini
+                if (UseProxyIni)
                 {
-                    Connected();
-                    Clientflag = true;
-                    /// Delete proxy.ini or restore settings.ini
-                    if (UseProxyIni)
-                    {
-                        if (File.Exists(FurcPath.GetInstallPath() + "proxy.ini"))
-                            File.Delete(FurcPath.GetInstallPath() + "proxy.ini");
-                    }
-                    /// reset settings.ini 10second delay timer
-                    NewsTimer = new System.Timers.Timer(10000);
-                    NewsTimer.Elapsed += new ElapsedEventHandler(OnTimedEvent);
-                    NewsTimer.Enabled = true;
-                    NewsTimer.AutoReset = false;
+                    if (File.Exists(FurcPath.GetInstallPath() + "proxy.ini"))
+                        File.Delete(FurcPath.GetInstallPath() + "proxy.ini");
                 }
+                // reset settings.ini 10second delay timer
+                NewsTimer = new System.Timers.Timer(10000);
+                NewsTimer.Elapsed += new ElapsedEventHandler(OnTimedEvent);
+                NewsTimer.Enabled = true;
+                NewsTimer.AutoReset = false;
             }
             catch (Exception e)
             {
@@ -716,17 +771,20 @@ namespace Furcadia.Net
                     //clientBuffer.Length = NetProxy.BUFFER_CAP
 
                     if (clientBuffer.Length <= 0)
-                        ClientDisConnected();
+                    {
+                        NetClientArgs Args = new NetClientArgs(Phase.ClientDisconnected);
+                        OnClientStatusChanged(this, Args);
+                    }
                     int pos = client.GetStream().Read(clientBuffer, 0, clientBuffer.Length);
                     clientBuild += System.Text.Encoding.GetEncoding(EncoderPage).GetString(clientBuffer, 0, pos);
                 }
 
-                //Every line should end with '\n'
+                //Every instruction ends with '\n'
                 //Split function removes the last character
                 lines.AddRange(clientBuild.Split('\n'));
                 for (int i = 0; i < lines.Count; i++)
                 {
-                    ClientData2(lines[i]);
+                    ClientData2(lines[i], System.EventArgs.Empty);
                     // we want ServerConnected and Check for null data Application may intentionally
                     // return ClientData = null to Avoid "Throat Tired" Syndrome. Let Application
                     // handle packet routing.
@@ -745,11 +803,18 @@ namespace Furcadia.Net
             }
             catch
             {// We don't care if there's an exception here.. We're Disconnecting any way
-                if (client.Connected == true) ClientDisConnected?.Invoke();
+                if (client.Connected == true)
+                {
+                    NetClientArgs Args = new NetClientArgs(Phase.Disconnected);
+                    OnClientStatusChanged(this, Args);
+                }
                 return;
             } // else throw e;
             if (IsClientConnected && clientBuild.Length < 1 || IsClientConnected == false)
-                ClientDisConnected?.Invoke();
+            {
+                NetClientArgs Args = new NetClientArgs(Phase.Disconnected);
+                OnClientStatusChanged(this, Args);
+            }
             if (client.Connected == true && clientBuild.Length >= 1)
             {
                 client.GetStream().BeginRead(clientBuffer, 0, clientBuffer.Length, new AsyncCallback(GetClientData), client);
@@ -771,7 +836,9 @@ namespace Furcadia.Net
 
                     if (read < 1)
                     {
-                        ServerDisConnected?.Invoke(); return;
+                        NetServerArgs Args = new NetServerArgs(Phase.Disconnected);
+                        OnServerStatusChanged(this, Args);
+                        return;
                     }
                     //If we have left over data add it to this server build
                     if (!string.IsNullOrEmpty(_ServerLeftOvers) && _ServerLeftOvers.Length > 0)
@@ -815,7 +882,8 @@ namespace Furcadia.Net
               // Detect if client disconnected
             if (IsServerConnected && serverBuild.Length < 1 || IsServerConnected == false)
             {
-                ServerDisConnected?.Invoke();
+                NetServerArgs Args = new NetServerArgs(Phase.Disconnected);
+                OnServerStatusChanged(this, Args);
             }
 
             if (IsServerConnected && serverBuild.Length > 0)
